Started restyling step selector for greater control over what user inputs

+++ app/forms/quantity_form.py
@@ -1,8 +1,8 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField
+from wtforms import StringField, IntegerField, FloatField
 from wtforms.validators import DataRequired
 
 class QuantityForm(FlaskForm):
     ingredient = StringField('ingredient', validators=[DataRequired()])
-    ingredient_quantity = IntegerField('ingredient_quantity', validators=[DataRequired()])
+    ingredient_quantity = FloatField('ingredient_quantity', validators=[DataRequired()])
     measurement_id = IntegerField('measurement_id', validators=[DataRequired()])
