+++ app/forms/steps_form.py
@@ -1,7 +1,7 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField
+from wtforms import StringField, IntegerField, FloatField
 from wtforms.validators import DataRequired
 
 class StepForm(FlaskForm):
-    step_number = IntegerField('step_number', validators=[DataRequired()])
+    step_number = FloatField('step_number', validators=[DataRequired()])
     step_description = StringField('step_description', validators=[DataRequired()])
